src/car_logo/requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
Pillow>=8.0.0
numpy>=1.20.0